requirements.txt
python-telegram-bot==7.0.1
python-graphql-client